requirements.txt
aiohttp                 
aiosignal                
anyio                     
async-timeout            
attrs                    
backoff                 
boto3                   
botocore                 
CacheControl             
cachetools               
certifi                
cffi                    
chardet                  
charset-normalizer                  
click                 
colorama              
coloredlogs            
cryptography          
dataclasses-json                      
Deprecated              
dill                    
entrypoints             
fastapi                
filelock                
filetype                 
firebase-admin           
flatbuffers             
frozenlist              
fsspec                  
gcloud                   
google-api-core          
google-api-python-client 
google-auth              
google-auth-httplib2     
google-cloud-core       
google-cloud-firestore   
google-cloud-storage    
google-crc32c          
google-resumable-media   
googleapis-common-protos 
greenlet                 
grpcio                  
grpcio-status           
gunicorn                
h11                     
httplib2                 
httptools                
huggingface-hub          
humanfriendly          
humbug                  
idna                   
importlib-resources     
Jinja2                   
jmespath                
joblib                  
jwcrypto                
langchain               
langchainplus-sdk       
lxml                     
MarkupSafe               
marshmallow              
marshmallow-enum         
monotonic                
mpmath                   
msgpack                  
multidict               
multiprocess            
mypy-extensions         
networkx                 
nltk                     
numcodecs               
numexpr                  
numpy                  
oauth2client            
onnxruntime            
openai                   
openapi-schema-pydantic 
overrides                
packaging                
pandas                   
pathos
pinecone-client                   
Pillow                  
pip                      
posthog                  
pox                     
ppft                    
proto-plus              
protobuf                 
pulsar-client            
pyasn1                  
pyasn1-modules           
pycparser               
pycryptodome             
pydantic                 
PyJWT                    
pyparsing               
PyPika                  
pyreadline3             
Pyrebase4               
python-dateutil          
python-dotenv           
python-jwt               
python-magic           
pytz                     
PyYAML                  
regex                    
requests                
requests-toolbelt       
rsa                    
s3transfer              
safetensors            
scikit-learn            
scipy                   
sentence-transformers    
sentencepiece            
setuptools               
six                     
sniffio                  
SQLAlchemy               
starlette               
sympy                    
tabulate                 
tenacity                 
threadpoolctl           
tiktoken                
tokenizers               
torch                    
torchvision              
tqdm                     
transformers             
typing_extensions        
typing-inspect           
tzdata                   
unstructured             
uritemplate              
urllib3                 
uvicorn                  
watchfiles              
websockets               
wrapt                    
yarl                    
